requirements.txt
streamlit==1.31.1
requests==2.31.0
python-dotenv==1.0.0
alibabacloud-oss-v2==2.22.0
pandas==2.2.0
numpy==1.26.3
openai==1.33.0